requirements.txt
requests
simplejson
six
lxml
xmltodict
python-weixin
sqlalchemy
psycopg2
phpserialize
flask
Jinja2
MarkupSafe
cython
pycryptodome
Pillow
Flask-RESTful
Flask-Cors
jsonschema
qiniu
pyjwt
celery
redis
Crypto
psycopg2-binary

